server/requirements.txt
asgiref==3.8.1
Django==5.0.6
django-cors-headers==4.4.0
djangorestframework==3.15.2
PyJWT==2.8.0
redis==5.0.7
sqlparse==0.5.0
tzdata==2024.1
